chore: migrate 10 APIs to owlbot
APIs:

RecommendationEngine
Recommender
Redis
ResourceManager
ResourceSettings
Retail
Scheduler
SecretManager
SecurityCenter
SecurityPrivateCa
