+++ Recommender/.OwlBot.yaml
@@ -1,0 +1,3 @@
+deep-copy-regex:
+    - source: /google/cloud/recommender/(v1)/.*-php/(.*)
+      dest: /owl-bot-staging/Recommender/$1/$2
