+++ Redis/.OwlBot.yaml
@@ -1,0 +1,3 @@
+deep-copy-regex:
+    - source: /google/cloud/redis/(v1|v1beta1)/.*-php/(.*)
+      dest: /owl-bot-staging/Redis/$1/$2
