+++ ResourceManager/owlbot.py
@@ -1,4 +1,4 @@
-# Copyright 2020 Google LLC
+# Copyright 2021 Google LLC
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
@@ -14,31 +14,25 @@
 
 """This script is used to synthesize generated parts of this library."""
 
-import synthtool as s
-import synthtool.gcp as gcp
 import logging
+from pathlib import Path
+import subprocess
+
+import synthtool as s
+from synthtool.languages import php
+from synthtool import _tracked_paths
 
 logging.basicConfig(level=logging.DEBUG)
 
-gapic = gcp.GAPICBazel()
-common = gcp.CommonTemplates()
+src = Path(f"../{php.STAGING_DIR}/ResourceManager").resolve()
+dest = Path().resolve()
 
-for version in ['v1', 'v1beta1']:
-    library = gapic.php_library(
-        service='security-private-ca',
-        version=version,
-        bazel_target=f'//google/cloud/security/privateca/{version}:google-cloud-security-privateca-{version}-php',
-    )
+# Added so that we can pass copy_excludes in the owlbot_main() call
+_tracked_paths.add(src)
 
-    # copy all src including partial veneer classes
-    s.move(library / 'src')
+php.owlbot_main(src=src, dest=dest)
 
-    # copy proto files to src also
-    s.move(library / 'proto/src/Google/Cloud/Security/PrivateCA', 'src/')
-    s.move(library / 'tests/')
 
-    # copy GPBMetadata file to metadata
-    s.move(library / 'proto/src/GPBMetadata/Google/Cloud/Security/Privateca', 'metadata/')
 
 # document and utilize apiEndpoint instead of serviceAddress
 s.replace(
@@ -57,24 +51,6 @@ s.replace(
     "**/Gapic/*GapicClient.php",
     r"\$transportConfig, and any \$serviceAddress",
     r"$transportConfig, and any `$apiEndpoint`")
-
-# fix year
-s.replace(
-    'src/V1beta1/**/*.php',
-    r'Copyright \d{4}',
-    r'Copyright 2020')
-s.replace(
-    'tests/*/V1beta1/*Test.php',
-    r'Copyright \d{4}',
-    r'Copyright 2020')
-s.replace(
-    'src/V1/**/*.php',
-    r'Copyright \d{4}',
-    r'Copyright 2021')
-s.replace(
-    'tests/*/V1/*Test.php',
-    r'Copyright \d{4}',
-    r'Copyright 2021')
 
 # Change the wording for the deprecation warning.
 s.replace(
