--- ResourceManager/synth.metadata
@@ -1,30 +1,0 @@
-{
-  "sources": [
-    {
-      "git": {
-        "name": ".",
-        "remote": "https://github.com/googleapis/google-cloud-php.git",
-        "sha": "a538ece434e64d4d8f03b7f2cfeb03482960daac"
-      }
-    },
-    {
-      "git": {
-        "name": "googleapis",
-        "remote": "https://github.com/googleapis/googleapis.git",
-        "sha": "4000ab10d80a78f2cd0690c2d0f347de50bc3d43",
-        "internalRef": "415061837"
-      }
-    }
-  ],
-  "destinations": [
-    {
-      "client": {
-        "source": "googleapis",
-        "apiName": "resourcemanager",
-        "apiVersion": "V3",
-        "language": "php",
-        "generator": "bazel"
-      }
-    }
-  ]
-}
