+++ ResourceSettings/.OwlBot.yaml
@@ -1,0 +1,3 @@
+deep-copy-regex:
+    - source: /google/cloud/resourcesettings/(v1)/.*-php/(.*)
+      dest: /owl-bot-staging/ResourceSettings/$1/$2
