+++ ResourceSettings/owlbot.py
@@ -1,4 +1,4 @@
-# Copyright 2019 Google LLC
+# Copyright 2021 Google LLC
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
@@ -14,32 +14,25 @@
 
 """This script is used to synthesize generated parts of this library."""
 
-import synthtool as s
-import synthtool.gcp as gcp
 import logging
+from pathlib import Path
+import subprocess
+
+import synthtool as s
+from synthtool.languages import php
+from synthtool import _tracked_paths
 
 logging.basicConfig(level=logging.DEBUG)
 
-gapic = gcp.GAPICBazel()
-common = gcp.CommonTemplates()
+src = Path(f"../{php.STAGING_DIR}/ResourceSettings").resolve()
+dest = Path().resolve()
 
-for version in ['V1', 'V1p1beta1']:
-    lower_version = version.lower()
-    library = gapic.php_library(
-        service='securitycenter',
-        version=lower_version,
-        bazel_target=f'//google/cloud/securitycenter/{lower_version}:google-cloud-securitycenter-{lower_version}-php',
-    )
+# Added so that we can pass copy_excludes in the owlbot_main() call
+_tracked_paths.add(src)
 
-    # copy all src
-    s.move(library / f'src/{version}')
+php.owlbot_main(src=src, dest=dest)
 
-    # copy proto files to src also
-    s.move(library / f'proto/src/Google/Cloud/SecurityCenter', f'src/')
-    s.move(library / f'tests/')
 
-    # copy GPBMetadata file to metadata
-    s.move(library / f'proto/src/GPBMetadata/Google/Cloud/Securitycenter', f'metadata/')
 
 # document and utilize apiEndpoint instead of serviceAddress
 s.replace(
@@ -59,36 +52,6 @@ s.replace(
     r"\$transportConfig, and any \$serviceAddress",
     r"$transportConfig, and any `$apiEndpoint`")
 
-# V1 is GA, so remove @experimental tags
-s.replace(
-    'src/V1/**/*Client.php',
-    r'^(\s+\*\n)?\s+\*\s@experimental\n',
-    '')
-
-# fix year
-s.replace(
-    'src/V1/**/*.php',
-    r'Copyright \d{4}',
-    r'Copyright 2019')
-s.replace(
-    'tests/*/V1/**/*Test.php',
-    r'Copyright \d{4}',
-    r'Copyright 2019')
-s.replace(
-    'src/V1p1beta1/**/*.php',
-    r'Copyright \d{4}',
-    r'Copyright 2020')
-s.replace(
-    'tests/*/V1p1beta1/**/*Test.php',
-    r'Copyright \d{4}',
-    r'Copyright 2020')
-
-# Use new namespace in the doc sample. See
-# https://github.com/googleapis/gapic-generator/issues/2141
-s.replace(
-    'src/V1/Gapic/SecurityCenterGapicClient.php',
-    r'Finding_State',
-    r'Finding\\State')
 # Change the wording for the deprecation warning.
 s.replace(
     'src/*/*_*.php',
@@ -126,3 +89,17 @@ s.replace(
     r"(.{0,})\]\((/.{0,})\)",
     r"\1](https://cloud.google.com\2)"
 )
+
+# format generated clients
+subprocess.run([
+    'npm',
+    'exec',
+    '--yes',
+    '--package=@prettier/plugin-php@^0.16',
+    '--',
+    'prettier',
+    '**/Gapic/*',
+    '--write',
+    '--parser=php',
+    '--single-quote',
+    '--print-width=80'])
