+++ Retail/.OwlBot.yaml
@@ -1,0 +1,3 @@
+deep-copy-regex:
+    - source: /google/cloud/retail/(v2)/.*-php/(.*)
+      dest: /owl-bot-staging/Retail/$1/$2
