--- Retail/synth.metadata
@@ -1,30 +1,0 @@
-{
-  "sources": [
-    {
-      "git": {
-        "name": ".",
-        "remote": "https://github.com/googleapis/google-cloud-php.git",
-        "sha": "5a619496b9f67caf1dda0194f04fd03a663a1b76"
-      }
-    },
-    {
-      "git": {
-        "name": "googleapis",
-        "remote": "https://github.com/googleapis/googleapis.git",
-        "sha": "cea896d1032b311179e5114b3e5a1b798e7040b7",
-        "internalRef": "387366941"
-      }
-    }
-  ],
-  "destinations": [
-    {
-      "client": {
-        "source": "googleapis",
-        "apiName": "retail",
-        "apiVersion": "v2",
-        "language": "php",
-        "generator": "bazel"
-      }
-    }
-  ]
-}
