+++ Scheduler/owlbot.py
@@ -1,4 +1,4 @@
-# Copyright 2018 Google LLC
+# Copyright 2019 Google LLC
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
@@ -14,48 +14,25 @@
 
 """This script is used to synthesize generated parts of this library."""
 
-import synthtool as s
-import synthtool.gcp as gcp
 import logging
+from pathlib import Path
+import subprocess
+
+import synthtool as s
+from synthtool.languages import php
+from synthtool import _tracked_paths
 
 logging.basicConfig(level=logging.DEBUG)
 
-gapic = gcp.GAPICBazel()
-common = gcp.CommonTemplates()
+src = Path(f"../{php.STAGING_DIR}/Scheduler").resolve()
+dest = Path().resolve()
 
-v1beta1_library = gapic.php_library(
-    service='redis',
-    version='v1beta1',
-    bazel_target='//google/cloud/redis/v1beta1:google-cloud-redis-v1beta1-php',
-)
+# Added so that we can pass copy_excludes in the owlbot_main() call
+_tracked_paths.add(src)
 
-# copy all src except partial veneer classes
-s.move(v1beta1_library / f'src/V1beta1/Gapic')
-s.move(v1beta1_library / f'src/V1beta1/resources')
+php.owlbot_main(src=src, dest=dest)
 
-# copy proto files to src also
-s.move(v1beta1_library / f'proto/src/Google/Cloud/Redis', f'src/')
-s.move(v1beta1_library / f'tests/')
 
-# copy GPBMetadata file to metadata
-s.move(v1beta1_library / f'proto/src/GPBMetadata/Google/Cloud/Redis', f'metadata/')
-
-v1_library = gapic.php_library(
-    service='redis',
-    version='v1',
-    bazel_target='//google/cloud/redis/v1:google-cloud-redis-v1-php',
-)
-
-# copy all src except partial veneer classes
-s.move(v1_library / f'src/V1/Gapic')
-s.move(v1_library / f'src/V1/resources')
-
-# copy proto files to src also
-s.move(v1_library / f'proto/src/Google/Cloud/Redis', f'src/')
-s.move(v1_library / f'tests/')
-
-# copy GPBMetadata file to metadata
-s.move(v1_library / f'proto/src/GPBMetadata/Google/Cloud/Redis', f'metadata/')
 
 # document and utilize apiEndpoint instead of serviceAddress
 s.replace(
@@ -81,43 +58,62 @@ s.replace(
     r'^(\s+\*\n)?\s+\*\s@experimental\n',
     '')
 
-
-# fix year
-s.replace(
-    '**/Gapic/*GapicClient.php',
-    r'Copyright \d{4}',
-    r'Copyright 2018')
-s.replace(
-    'tests/**/V1*/*Test.php',
-    r'Copyright \d{4}',
-    r'Copyright 2018')
-
 # Change the wording for the deprecation warning.
 s.replace(
     'src/*/*_*.php',
     r'will be removed in the next major release',
     'will be removed in a future release')
 
-# Fix class references in gapic samples
-for version in ['V1', 'V1beta1']:
-    pathExpr = 'src/' + version + '/Gapic/CloudRedisGapicClient.php'
+# Fix missing formatting method
+s.replace(
+    'src/V1/Gapic/CloudSchedulerGapicClient.php',
+    r"private static \$jobNameTemplate;\n\s{4}private static \$locationNameTemplate;",
+    """private static $jobNameTemplate;
+    private static $locationNameTemplate;
+    private static $projectNameTemplate;"""
+)
+s.replace(
+    'src/V1/Gapic/CloudSchedulerGapicClient.php',
+    "private static function getPathTemplateMap",
+    """private static function getProjectNameTemplate()
+    {
+        if (null == self::$projectNameTemplate) {
+            self::$projectNameTemplate = new PathTemplate('projects/{project}');
+        }
 
-    types = {
-        'new CloudRedisClient': r'new Google\\Cloud\\Redis\\'+ version + r'\\CloudRedisClient',
-        'new Instance': r'new Google\\Cloud\\Redis\\' + version + r'\\Instance',
-        '= Tier::': r'= Google\\Cloud\\Redis\\' + version + r'\\Instance\\Tier::',
-        'new FieldMask': r'new Google\\Protobuf\\FieldMask',
-        'new InputConfig': r'new Google\\Cloud\\Redis\\' + version + r'\\InputConfig',
-        'new OutputConfig': r'new Google\\Cloud\\Redis\\' + version + r'\\OutputConfig',
-        '= DataProtectionMode': r'= Google\\Cloud\\Redis\\' + version + r'\\FailoverInstanceRequest\\DataProtectionMode::'
+        return self::$projectNameTemplate;
     }
 
-    for search, replace in types.items():
-        s.replace(
-            pathExpr,
-            search,
-            replace
-        )
+    private static function getPathTemplateMap"""
+)
+s.replace(
+    'src/V1/Gapic/CloudSchedulerGapicClient.php',
+    r"'job' => self::getJobNameTemplate\(\),\n\s{0,}'location' => self::getLocationNameTemplate\(\),",
+    """'job' => self::getJobNameTemplate(),
+                'location' => self::getLocationNameTemplate(),
+                'project' => self::getProjectNameTemplate(),"""
+)
+s.replace(
+    'src/V1/Gapic/CloudSchedulerGapicClient.php',
+    r"\/\*\*\n\s{5}\* Parses a formatted name string and returns an",
+    """/**
+     * Formats a string containing the fully-qualified path to represent
+     * a project resource.
+     *
+     * @param string $project
+     *
+     * @return string The formatted project resource.
+     */
+    public static function projectName($project)
+    {
+        return self::getProjectNameTemplate()->render([
+            'project' => $project,
+        ]);
+    }
+
+    /**
+     * Parses a formatted name string and returns an"""
+)
 
 ### [START] protoc backwards compatibility fixes
