--- SecretManager/synth.metadata
@@ -1,39 +1,0 @@
-{
-  "sources": [
-    {
-      "git": {
-        "name": ".",
-        "remote": "https://github.com/googleapis/google-cloud-php.git",
-        "sha": "9ca2e74c0f33250c1547a5d2fa47cced78c1a5df"
-      }
-    },
-    {
-      "git": {
-        "name": "googleapis",
-        "remote": "https://github.com/googleapis/googleapis.git",
-        "sha": "4000ab10d80a78f2cd0690c2d0f347de50bc3d43",
-        "internalRef": "415061837"
-      }
-    }
-  ],
-  "destinations": [
-    {
-      "client": {
-        "source": "googleapis",
-        "apiName": "secretmanager",
-        "apiVersion": "v1",
-        "language": "php",
-        "generator": "bazel"
-      }
-    },
-    {
-      "client": {
-        "source": "googleapis",
-        "apiName": "secrets",
-        "apiVersion": "v1beta1",
-        "language": "php",
-        "generator": "bazel"
-      }
-    }
-  ]
-}
