--- SecurityCenter/synth.metadata
@@ -1,39 +1,0 @@
-{
-  "sources": [
-    {
-      "git": {
-        "name": ".",
-        "remote": "https://github.com/googleapis/google-cloud-php.git",
-        "sha": "8ac34686059df4f10244db3b24abe4439273605b"
-      }
-    },
-    {
-      "git": {
-        "name": "googleapis",
-        "remote": "https://github.com/googleapis/googleapis.git",
-        "sha": "be8f9889c7662043b89f63d475e72eea9a0621af",
-        "internalRef": "411093163"
-      }
-    }
-  ],
-  "destinations": [
-    {
-      "client": {
-        "source": "googleapis",
-        "apiName": "securitycenter",
-        "apiVersion": "v1",
-        "language": "php",
-        "generator": "bazel"
-      }
-    },
-    {
-      "client": {
-        "source": "googleapis",
-        "apiName": "securitycenter",
-        "apiVersion": "v1p1beta1",
-        "language": "php",
-        "generator": "bazel"
-      }
-    }
-  ]
-}
