+++ SecurityPrivateCa/.OwlBot.yaml
@@ -1,0 +1,3 @@
+deep-copy-regex:
+    - source: /google/cloud/security/privateca/(.*)/.*-php/(.*)
+      dest: /owl-bot-staging/SecurityPrivateCa/$1/$2
