--- SecurityPrivateCa/synth.metadata
@@ -1,46 +1,0 @@
-{
-  "sources": [
-    {
-      "git": {
-        "name": ".",
-        "remote": "https://github.com/googleapis/google-cloud-php",
-        "sha": "fca983c6cf8b7802cf31fb4f952fad09dc2617f9"
-      }
-    },
-    {
-      "git": {
-        "name": "googleapis",
-        "remote": "https://github.com/googleapis/googleapis.git",
-        "sha": "e9f5f8d2f25146730e80574778602aca38a70823",
-        "internalRef": "382547062"
-      }
-    },
-    {
-      "git": {
-        "name": "synthtool",
-        "remote": "https://github.com/googleapis/synthtool.git",
-        "sha": "39652e3948f455fd0b77535a0145eeec561a3706"
-      }
-    }
-  ],
-  "destinations": [
-    {
-      "client": {
-        "source": "googleapis",
-        "apiName": "security-private-ca",
-        "apiVersion": "v1",
-        "language": "php",
-        "generator": "bazel"
-      }
-    },
-    {
-      "client": {
-        "source": "googleapis",
-        "apiName": "security-private-ca",
-        "apiVersion": "v1beta1",
-        "language": "php",
-        "generator": "bazel"
-      }
-    }
-  ]
-}
